+++ packages/ui/src/icons/SparkleIcon.tsx
